% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing.R
\name{diversity}
\alias{diversity}
\title{Fishing function
This function takes a vector of fishing data and calculates the most common species, rarest species
and the total number of species.}
\usage{
diversity(species)
}
\arguments{
\item{species}{is a vector of species}
}
\value{
list with the following items
\describe{
\item{most_commom_species}{Name of the most common species}
\item{rerest_species}{Name of the rarest species}
\item{total_species}{Total number of species}}
}
\description{
Fishing function
This function takes a vector of fishing data and calculates the most common species, rarest species
and the total number of species.
}
