% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcount.R
\docType{data}
\name{newcount}
\alias{newcount}
\title{Sun data for plant growth}
\format{
A vector of sun data
\itemize{
 \item data for number}
}
\usage{
newcount
}
\description{
Data for plant_growth_function for random number of weekly hours each week for a year
}
\keyword{datasets}
